pyproject.toml
[project]
name = "automated-video-description-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.84.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
    "typer>=0.16.0",
]

[project.scripts]
video-describe = "main:main"
